Moved all ChoiceFieldTests to non-abstract classes
